baseline_json_editor/pyproject.toml
[project]
name = "baseline-json-editor"
version = "0.1.0"
description = "Convert documents (PDF, DOCX, etc.) to an editable baseline JSON schema using Marker, with a premium Streamlit UI."
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "marker-pdf>=0.3.0",
    "Pillow>=10.0.0",
    "pypdfium2>=4.20.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.40.0",
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
    "streamlit-ace>=0.1.1",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "mammoth>=1.9.0",
]

[project.optional-dependencies]
full = ["openpyxl>=3.1.5", "python-pptx>=1.0.2", "ebooklib>=0.18"]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]

[project.scripts]
baseline-editor = "app:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
